+++ web/app/components/workflow/__tests__/selection-contextmenu.spec.tsx
@@ -3,16 +3,12 @@ import { act, fireEvent, screen, waitFor } from '@testing-library/react'
 import { useEffect } from 'react'
 import { useNodes } from 'reactflow'
 import SelectionContextmenu from '../selection-contextmenu'
-import { AlignType } from '../selection-contextmenu.helpers'
 import { useWorkflowHistoryStore } from '../workflow-history-store'
 import { createEdge, createNode } from './fixtures'
 import { renderWorkflowFlowComponent } from './workflow-test-env'
 
 let latestNodes: Node[] = []
 let latestHistoryEvent: string | undefined
-const mockGetAlignBounds = vi.fn()
-const mockAlignNodePosition = vi.fn()
-const mockGetAlignableNodes = vi.fn()
 const mockGetNodesReadOnly = vi.fn()
 
 vi.mock('../hooks', async () => {
@@ -22,28 +18,6 @@ vi.mock('../hooks', async () => {
     useNodesReadOnly: () => ({
       getNodesReadOnly: mockGetNodesReadOnly,
     }),
-  }
-})
-
-vi.mock('../selection-contextmenu.helpers', async () => {
-  const actual = await vi.importActual<typeof import('../selection-contextmenu.helpers')>('../selection-contextmenu.helpers')
-  return {
-    ...actual,
-    alignNodePosition: (...args: Parameters<typeof actual.alignNodePosition>) => {
-      if (mockAlignNodePosition.getMockImplementation())
-        return mockAlignNodePosition(...args)
-      return actual.alignNodePosition(...args)
-    },
-    getAlignableNodes: (...args: Parameters<typeof actual.getAlignableNodes>) => {
-      if (mockGetAlignableNodes.getMockImplementation())
-        return mockGetAlignableNodes(...args)
-      return actual.getAlignableNodes(...args)
-    },
-    getAlignBounds: (...args: Parameters<typeof actual.getAlignBounds>) => {
-      if (mockGetAlignBounds.getMockImplementation())
-        return mockGetAlignBounds(...args)
-      return actual.getAlignBounds(...args)
-    },
   }
 })
 
@@ -97,9 +71,6 @@ describe('SelectionContextmenu', () => {
     vi.clearAllMocks()
     latestNodes = []
     latestHistoryEvent = undefined
-    mockGetAlignBounds.mockReset()
-    mockAlignNodePosition.mockReset()
-    mockGetAlignableNodes.mockReset()
     mockGetNodesReadOnly.mockReset()
     mockGetNodesReadOnly.mockReturnValue(false)
   })
@@ -241,10 +212,9 @@ describe('SelectionContextmenu', () => {
   })
 
   it('should cancel when align bounds cannot be resolved', () => {
-    mockGetAlignBounds.mockReturnValue(null)
     const nodes = [
-      createNode({ id: 'n1', selected: true, width: 40, height: 20 }),
-      createNode({ id: 'n2', selected: true, position: { x: 80, y: 20 }, width: 40, height: 20 }),
+      createNode({ id: 'n1', selected: true }),
+      createNode({ id: 'n2', selected: true, position: { x: 80, y: 20 } }),
     ]
 
     const { store } = renderSelectionMenu({ nodes })
@@ -280,10 +250,15 @@ describe('SelectionContextmenu', () => {
 
   it('should cancel when alignable nodes shrink to one item', () => {
     const nodes = [
-      createNode({ id: 'n1', selected: true, width: 40, height: 20 }),
-      createNode({ id: 'n2', selected: true, position: { x: 80, y: 20 }, width: 40, height: 20 }),
+      createNode({
+        id: 'container',
+        selected: true,
+        width: 40,
+        height: 20,
+        data: { _children: [{ nodeId: 'child', nodeType: 'code' as never }] },
+      }),
+      createNode({ id: 'child', selected: true, position: { x: 80, y: 20 }, width: 40, height: 20 }),
     ]
-    mockGetAlignableNodes.mockImplementation((allNodes: Node[]) => [allNodes[0]])
 
     const { store } = renderSelectionMenu({ nodes })
 
@@ -294,29 +269,7 @@ describe('SelectionContextmenu', () => {
     fireEvent.click(screen.getByTestId('selection-contextmenu-item-left'))
 
     expect(store.getState().selectionMenu).toBeUndefined()
-    expect(latestNodes.find(node => node.id === 'n1')?.position.x).toBe(0)
-    expect(latestNodes.find(node => node.id === 'n2')?.position.x).toBe(80)
-  })
-
-  it('should skip align updates when a selected node is not found in the draft', () => {
-    const nodes = [
-      createNode({ id: 'n1', selected: true, position: { x: 20, y: 40 }, width: 40, height: 20 }),
-      createNode({ id: 'n2', selected: true, position: { x: 140, y: 90 }, width: 60, height: 30 }),
-    ]
-    mockGetAlignableNodes.mockImplementation((allNodes: Node[]) => [
-      allNodes[0],
-      { ...allNodes[1], id: 'missing-node' },
-    ])
-
-    const { store } = renderSelectionMenu({ nodes })
-
-    act(() => {
-      store.setState({ selectionMenu: { left: 100, top: 100 } })
-    })
-
-    fireEvent.click(screen.getByTestId(`selection-contextmenu-item-${AlignType.Right}`))
-
-    expect(latestNodes.find(node => node.id === 'n1')?.position.x).toBe(160)
-    expect(latestNodes.find(node => node.id === 'n2')?.position.x).toBe(140)
+    expect(latestNodes.find(node => node.id === 'container')?.position.x).toBe(0)
+    expect(latestNodes.find(node => node.id === 'child')?.position.x).toBe(80)
   })
 })
